Code cleanup (2)
