+++ api-generator/generator-interfaces/src/main/java/esa/mo/tools/stubgen/writers/InterfaceMethodBuilder.java
@@ -1,5 +1,5 @@
 /* ----------------------------------------------------------------------------
- * Copyright (C) 2013      European Space Agency
+ * Copyright (C) 2026      European Space Agency
  *                         European Space Operations Centre
  *                         Darmstadt
  *                         Germany
